requirements.txt
urllib3==1.24.2
requests==2.22.0
lxml==4.4.1
xlwt==1.3.0
python_dateutil==2.8.1
